requirements.txt
eventlet==0.25.1
requests==2.22.0
tinydb==3.15.2
pyyaml==5.3.1
